Move Domains struct & static URL methods to separate extension file
