+++ UniHub/UniHub/UniAPI/APIClient.swift
@@ -57,26 +57,3 @@ class APIClient {
         task.resume()
     }
 }
-
-extension APIClient {
-    private struct Domains {
-        static let baseURL = "https://raw.githubusercontent.com/Hipo/university-domains-list/master/world_universities_and_domains.json"
-        static let testURL = "http://localhost:8080"
-    }
-    
-    func getAllUniversitiesURL() -> URL {
-        guard let fullURL = URL(string: Domains.baseURL) else {
-            let failureURL = URL(string: "")
-            return failureURL!
-        }
-        return fullURL
-    }
-    
-    func getTestURL() -> URL {
-        guard let fullURL = URL(string: Domains.testURL) else {
-            let failureURL = URL(string: "")
-            return failureURL!
-        }
-        return fullURL
-    }
-}
